move hash calculation tests out and fix them
